[Refactor] #428 입력 컴포넌트 수정
