+++ app/src/main/java/kr/co/lion/modigm/ui/join/JoinStep1EmailAndPwScreen.kt
@@ -1,9 +1,10 @@
-package kr.co.lion.modigm.ui.join.component
+package kr.co.lion.modigm.ui.join
 
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.foundation.layout.padding
 import androidx.compose.foundation.rememberScrollState
+import androidx.compose.foundation.text.KeyboardOptions
 import androidx.compose.foundation.verticalScroll
 import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
@@ -16,6 +17,8 @@ import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
 import androidx.lifecycle.compose.collectAsStateWithLifecycle
 import kr.co.lion.modigm.R
+import kr.co.lion.modigm.ui.join.component.EmailTextInputField
+import kr.co.lion.modigm.ui.join.component.PasswordTextInputField
 import kr.co.lion.modigm.ui.join.vm.JoinStep1EmailAndPwViewModel
 
 @Composable
@@ -48,7 +51,7 @@ fun JoinStep1EmailAndPwScreen(
             onValueChange = { joinStep1EmailAndPwViewModel.setUserInputEmail(it) },
             isError = emailValidationMessageState.value.isNotEmpty(),
             errorMessage = emailValidationMessageState.value,
-            imeAction = ImeAction.Next,
+            keyboardOptions = KeyboardOptions(imeAction = ImeAction.Next),
             modifier = Modifier.padding(top = 40.dp)
         )
         PasswordTextInputField(
@@ -57,7 +60,7 @@ fun JoinStep1EmailAndPwScreen(
             onValueChange = { joinStep1EmailAndPwViewModel.setUserInputPassword(it) },
             isError = passwordValidationMessageState.value.isNotEmpty(),
             errorMessage = passwordValidationMessageState.value,
-            imeAction = ImeAction.Next,
+            keyboardOptions = KeyboardOptions(imeAction = ImeAction.Next),
             modifier = Modifier.padding(top = 16.dp)
         )
         Text(
@@ -71,7 +74,7 @@ fun JoinStep1EmailAndPwScreen(
             onValueChange = { joinStep1EmailAndPwViewModel.setUserInputPasswordCheck(it) },
             isError = passwordCheckValidationMessageState.value.isNotEmpty(),
             errorMessage = passwordCheckValidationMessageState.value,
-            imeAction = ImeAction.Done,
+            keyboardOptions = KeyboardOptions(imeAction = ImeAction.Done),
             modifier = Modifier.padding(top = 16.dp)
         )
     }
